refactor -> move ButtonView,InputField,RoomButton to interactive package
