+++ ChatterClient/src/main/java/com/codecool/chatter/view/CreateRoomForm.java
@@ -1,5 +1,7 @@
 package com.codecool.chatter.view;
 
+import com.codecool.chatter.view.interactive.ButtonView;
+import com.codecool.chatter.view.interactive.InputField;
 import javafx.event.EventHandler;
 import javafx.geometry.Insets;
 import javafx.scene.input.MouseEvent;
