+++ ChatterClient/src/main/java/com/codecool/chatter/view/interactive/InputField.java
@@ -1,4 +1,4 @@
-package com.codecool.chatter.view;
+package com.codecool.chatter.view.interactive;
 
 import javafx.geometry.HPos;
 import javafx.geometry.Insets;
